buffer: implement TILE_COPY in the cache handler
If the tile exists in the cache, we start by copying it directly to
the destination.  Then, if the tile doesn't exist in the cache, or
if it's already stored -- i.e., if it might exist in the backend in
an up-to-date state -- we try letting the backend copy the tile as
well, and, if this succeeds, we either mark the tile copied from
the cache as stored, or void any existing tile in the destination
cache.  The copy is considered successful if either the cache or
the backend copied the tile.