buffer, stats: implement TILE_COPY in the swap backend; cleanup; new stat
In GeglTileBackendSwap, implement the TILE_COPY command between a
pair of swap backends, but decoupling swap data blocks,
representing an allocated block in the common swap file, from swap
entries, representing stored tiles in an individual swap backend.
Have data blocks be ref-counted, such that each swap entry
corresponds to a single data block, but each data block may be
shared among multiple entries.  When copying a tile, we simply
assign the data block of the source entry to the destination entry,
and increment its reference count.  When modifying a tile, upon a
TILE_SET command, if the corresponding data block isn't uniquely
owned by the entry (i.e., if its ref-count is greater than 1), we
allocate a new data block for the tile.

Some all-around cleanup in gegl-tile-backend-swap.c.

In GeglStats, add a "swap-total-uncloned" stat, analogous to
"cache-total-uncloned", reporting the total size of the data that
would have been stored in the swap, if each tile occupied a unique
data block.