buffer, stats: limit maximal pending data size in the swap backend queue
In GeglTileBackendSwap, limit the maximal data size allowed to be
pending in the queue at any given time.  The limit is currently set
to 10% of the cache size.  If data is attempted to be written to
the swap while the queue is full, the operation blocks until the
queued data size drops below the limit.

Add new "swap-queue-total", "swap-queue-full", and "swap-queue-
stalls" properties to GeglStats, reporting the total queued data
size, whether or not the queue is full, and the number of times a
write operation has been blocked due to a full queue, respectively.