long-shadow: add midpoint parameter to finite fading styles
Add a "midpoint-rel" property, which acts as the fade midpoint
parameter, instead of the "midpoint" property, for the fadinig-
fixed-length and fading-fixed-rate styles.  Unlike the latter,
"midpoint-rel" is specified as a factor of the shadow length,
instead of as an absolute distance.  When using GUM, as GIMP does,
the label of both properties is "midpoint", so they appear to be a
single property.

The midpoint value determines the gamma curve to use as the drop-
off function of the shadow.  When it is 0.5, the drop-off is
linear, as before; when it is < 0.5, the drop-off decelerates; and,
when it is > 0.5, the drop-off accelerates.

Note that in the fixed-rate style, higher midpoint values result in
darker but *shorter* shadows, and a 1.0 midpoint results in no
shadow at all.