buffer,opencl: split opencl code from GeglBuffer
The integration between the opencl buffer cache and GeglBuffer consists of
two (three) function calls that are conditionally done for updates are
correctly propagated between opencl / cpu.

This commit in work towards making GeglBuffer possible to reuse as a separate
libraryt without needing to link and load opencl replaces these conditionals
with a check whether a function pointer is set, and if it is set it uses it.
The opencl code sets/unsets the function pointers depending on whether it is
active.