buffer: add gegl_buffer_flush_ext
This is the call to be used externally where the explicit opencl cache
flushing was used before. In code that uses opencl existing checks for
opencl can be used around it - but do not have to, since it is a no-op
with no flush_ext vfuncs set, like it is with no opencl in operation.