buffer: mark old gegl_buffer_iterator_new function as deprecated
The way to get rid of deprecation warnings is to, at the top of the file before
gegl-buffer-iterator2.h the 2 suffix is not to be used, effectively this means
that the constructor now takes an additional argument which is the maximum
number of buffers the iterator should be allocated to support, and access to
roi/data members need to be adjusted to be fields of the new array items.

When we've gotten rid of all warnings in GIMP the following release of GEGL,
gegl-buffer-iterator2.h will be copied over gegl-buffer-iterator.h and the
macro will no longer add the 2 suffix.