Issue #97 - Drop Shadow crashes GIMP if opacity is adjusted to 1.000 ...
... with the GUI arrow buttons

In gegl:opacity, use fuzzy comparisson for checking if the "value"
property equals 1, and the op is a pass-through.

This test was previously performed using exact comparison, in a
context where "value" is a double, and then later asserted-on in a
context where "value" is a float, which could lead to a different
result, aborting the program.  This commit performs all such tests
with "value" as a float, and removes the assertion.