buffer, tests: specify iterator aliasing rules, and prevent deadlocks
Specify the rules under which aliased buffers, i.e., buffers
sharing the same tile storage, may be added to the same iterator.
The rules state that if a buffer is acessed for writing, then the
iterated-over areas of each of its aliases must either overlap its
iterated-over area completely, or not at all, and that none of the
other overlapping aliases may be accessed for writing as well.

Make sure that when these rules are followed, no deadlocks occur
during iteration, regardless of the order in which the buffers are
added to the iterator.  This is done by locking all the write-
accessed tiles before the read-accessed tiles, and unlocking all
the read-accessed tiles before unlocking/gegl_buffer_set()ing the
write-accessed tiles.

Add a test case, checking that no deadlocks occur.