long-shadow: disable fixed-rate fading style
The fixed-rate fading shadow style's interaction with the midpoint
parameter is problematic, having a discontinuity at midpoint=1.  It
may also be problematic w.r.t. some future directions for the op.
Let's disable it for now, especially since the fixed-length fading
style is probably the more useful one.

Note that the code for the fixed-rate style is all still there,
it's simply made inaccessible.