operations: improve gegl:watershed-transform description.
It is completely unobvious how to use the operation without looking at
the code (in particular the fact that it uses the alpha channel to
detect what is labelled or not, and also the babl formats it will work
on since you likely don't want your labels converted). Let's add a
slightly more descriptive text.