buffer: fix 32bit integer boxfilter overflow
In the existing implementation we overflow due to the imprecision introduced by
using single precision floating point in the template. We introduce a temp type
that can be used for computations, as well as do an upper bounds check when
doing the rounding/converting of the final result. This fixes overflows but not
the imprecision; that only applies for 32bit integer precision and in code
paths primarily used for display rather than processing.

Fixing issue #103