buffer: apply gamma curve to alpha component during filtering
Disable the _alpha() variant of
gegl_{boxfilter,bilinear,downscale_2x2}_u8_nl(), and use the non-
alpha one instead, applying the same gamma curve to the alpha
components as we do for the color components.  This is a temporary
hack, meant to avoid producing over-saturated pixels when using
these functions with premultiplied-alpha formats (in particular,
cairo-ARGB32), which can lead, most notably, to visual artifacts in
cairo during compositing (see issue #104).