Partially revert "buffer: fix source rect passed to gegl_resample_{linear,boxfilter}()"
Commit a7c730cf6e841a46d3b4546b86a234ae59206e9c trimmed the sample-
buffer padding from the source-rect passed to the resampling
functions.  This isn't strictly necessary, and has the negative
effect of preventing the same padding from being present in the
converted sample-buffer when using any of the _generic() resampling
functions, causing artifacts.

Revert the trimming of the resampling source-rect, though keep
avoiding reading padding-region data from the source buffer, which
should still be unnecessary.