operations: watershed-transform: do not queue labels with no...
... unlabelled neighbours.
Previous algorithm was queuing any labelled (i.e. unflagged) pixel to
later unqueue it, check for its neighbour and do nothing when no
neighbours were unlabelled (flagged). This was very inefficient
especially as we were reading the buffer with gegl_buffer_get() when
unqueuing (buffer iterators are not usable then since we unqueue in
priority order).

Before queuing a label, let's make sure it has at least one unlabelled
neighbour, hence that it will result in actual transformation. On test
FullHD image on my computer, previous code was running in 1+ second,
while the optimized operation runs in about 0.1 second for the same
image (and of course, the result is unchanged, as this doesn't change
the actual logics).