operations: make generated point composers handle more pixel formats
Make the innerloops able to work with any number of components with or without
alpha, a possible performance enhancement would be to implement separate code
paths for formats with and without alpha, while still keeping the number of
components dynamic.

This makes the math and generated operations have code paths dealing with
grayscale and grayscale with alpha versions - reducing the needed memory
bandwidth, while also preparing for spectral or other multi-channel uses with
or without alpha channels.