warp: parallelize using gegl-parallel
gegl:warp can't use auto-threading.  Use gegl-parallel to
parallelize it manually.

Convert the op to C++ in order to use lamdas, which simplify
parallelizarion.  Move the op to a new common-cxx/ module, where we
can add C++ operations, and which is only built when a suitable C++
compiler is found.  We don't leave it in common/, since Automake
seems to always use CXXLD to link libraries that have (potential)
C++ sources, even if they're if-ed out after configuration.