map-{absolute,relative}: factor out common code; add scale matrix
Factor out the common code of gegl:map-absolute and
gegl:map-relative to map-common.h, and include it in both
operations.

When using a non-nearest sampler, use the gradient of the input
field to construct a scale matrix to pass to the sampler, to reduce
aliasing when the input is downscaled.