buffer: improve linear/cubic samplers box-filter criterion
When the linear and cubic samplers are given a scale matrix,
perform box filtering whenever the norm of any of the transformed
basis vectors is greater-than or equal-to 2, that is, roughly
speaking, whenever the image is scaled down by a factor of 2 or
more in *either* direction.  This replaces the old criterion,
checking whether the absolute value of the scale matrix's
determinent, corresponding to the sampled area, is greater-than or
equal-to 4, that is, roughly speaking, whether the image is scaled
down by a factor of 2 or more in *both* directions.

This is especially important for GIMP's warp tool (which, since
last commit, uses a scale matrix), since it very commonly produces
uneven scaling.