buffer: use calling sampler for sub-sampling during box filtering
When performing box filtering, in the linear and cubic samplers,
use the calling sampler to sub-sample the buffer, instead of using
a separate sampler of a different type.  This allows us to perform
box filtering even when downscaling only along one direction,
without loss of quality or discontinuity, undoing
commit 662b9c35f5e052f94268ea0aea59923ed23b10bb.

Note that this can make box filtering slower, since we used to use
a cheaper sampler for sub-sampling.  This is somewhat countered by
reducing the number of sub-samples.