voronoi-diagram: new op in workshop
Given a subset of the input referred to as the "seed", the
gegl:voronoi-diagram op paints each non-seed pixel with the color
of the nearest seed pixel, using one of several metrics.  The non-
seeded region is identified either through a special color in the
input, or using an AUX input.  The resulting output is, essentialy,
the Voronoi diagram of the input.