buffer: damage tile pyramid when voiding uncached base tiles
In GeglTileHandlerCache, when trying to void a base tile that's not
present in the cache, go ahead and void the cooresponding area of
the tile pyramid anyway, since it might include cached mipmapped
tiles.  In particular, this may happen if the corresponding area of
the buffer is empty, but mipmapped tiles have been generated for it
(which doesn't cause corresponding empty base tiles to be created),
or if the corresponding base tiles have been swapped out.