buffer: don't damage pyramid when inserting new tile to cache
In gegl_tile_handler_cache_insert(), when there's already a cached
tile at the same coordinates as the new tile, merely remove the old
tile from the cache, but don't damage the corresponding area of the
pyramid.  We already explicitly damage the pyramid at all the
relevant points, so doing this again in cache_insert() is
superfluous.  Moreover, the code in gegl_tile_handler_cache_void()
that is responsible for the damaging is erroneous, and fixing it
(as done in the next few commits) would make damaging the pyramid
in cache_insert() excessive, breaking a few tests.

	.mtsync.log