buffer: avoid unnecesarily locking global cache mutex when inserting tile
In gegl_tile_handler_cache_insert(), only call
gegl_tile_handler_cache_trim() if the cache actually needs to be
trimmed, i.e., if the total cache size exceeds the maximum.  The
latter function always acquires the global cache mutex, which is
wasteful if there is nothing to actually trim.