workshop: add a gray component replacement implementation
This implementation assumes that C=M=Y is gray, this could be
improved by extracting some more information from the ICC profile
of the space. However the changes when converted to RGB for display
by GEGL using applications will be wysiwyg.

For now the inklimit slider has to be pushed to the right for the
final render. It might well be better to have two separate ops
for this and juggle them in ui code instead.