buffer: lower cost of cloned tiles queued for writing to the swap
In the swap tile backend, when calculating the cost of the tiles
queued for writing, divide the tile size by the number of clones it
has, so that if all clones of a given tile are queued for writing,
their total cost (i.e., the amount of memory they consume) is equal
to the size of a single tile.  This allows more cloned tiles to be
queued for writing simultaneously, reducing the number of stalls.