buffer, tests, perf: add gegl-compression
gegl-compression is a simple compression framework, meant for
(losslessly) compressing image data.  It exposes a set of
compression algorithms, which can be used to compress/decompress
data.  We're going to use gegl-compression to compress tile data
stored in the swap.

Add correctness and performance tests, which test all the available
algorithms.