gegl-compression: add RLE algorithms
Add a set of RLE compression algorithms: "rle1", "rle2", "rle4",
and "rle8".  The numeric suffix specifies the unit of compression,
in bits -- the algorithms independently compress the respective
units of each pixel.  That is, "rle1" RLE-compresses the first bit
of each pixel, then the second bit of each pixel, and so on, while
"rle8" RLE-compresses the first byte of each pixel, then the second
byte of each pixel, and so on.  Smaller units yield better, but
slower, compression.