GeglNode: Don't copy if the result is already in the destination
This can happen if the GeglNode has been explicitly asked to render to
its own GeglCache, even when it caches naturally during the processing
of the graph. eg., if the GeglNode and its underlying GeglOperation
haven't disabled caching and the area being rendered lies within the
node's bounding box.

In such cases, the GeglBuffer obtained out of processing the GeglNode
is its own GeglCache. Hence, there's no need to copy it to the
destination buffer if the destination is the cache itself.

https://gitlab.gnome.org/GNOME/gegl/merge_requests/14