processor: Simplify and optimize rendering to the cache
A GeglNode which is not a sink or is a sink that needs its entire input
in one go, has to be rendered to its own GeglCache. In such cases,
it is easier to ask gegl_node_blit to render to the cache, instead of
doing it manually by rendering to a linear memory buffer and then
writing it to the cache.

If the GeglNode caches naturally during processing, then there's no
need to separately write to the cache once the processing is over. This
is something that gegl_node_blit is smart enough to know.