buffer: in the swap backend, compress queued tiles while stalling
When the write queue is full while storing a tile to the swap
backend, walk the queue and compress any uncompressed tiles before
blocking the current thread.  Previously, we would only compress
the currently-stored tile in this situation.  This allows the
storing thread to participate in tile compression, reducing the
cost of the write queue, shortening stall time.