buffer: allow multiple threads to trim the cache concurrently
In the cache tile-handler, allow multiple threads to trim the cache
concurrently.  This is particularly useful since last commit, as it
allows all threads waiting on the cache to participate in swap tile
compression.

Note that each thread can only trim a separate cache tile-handler,
which means that some threads may trim tiles not belonging to the
LRU cache.  Hopefully, this wouldn't have a significant impact.