buffer: don't copy/zero data when uncloning fully-damaged tile
When uncloning a fully-damaged tile, allocate a new buffer for the
tile data, but don't copy the old data over (or zero the new
buffer, if it's an empty tile), since the existing data is
conceptually invalid, and will be overwritten anyway.

The next few commits fully damage tiles that are going to be fully
overwritten, so that, if they have to be uncloned, we don't
needlessly copy/zero their data.