buffer: add GeglTile::keep_identity flag
Add a keep_identity flag to GeglTile, which indicates that the
tile's data pointer's identity must be maintained.  When the flag
is set, we avoid cloning the tile in gegl_tile_dup(), and rather
perform a direct copy, since we have to way to unclone the tile
without changing its data pointer.  Additionally, we avoid trimming
the tile out of the cache, since the tile storage might copy its
data and get rid of the tile, returning a new tile, with a
different data pointer, when re-fetching it.