buffer: fix race condition when disconnecting cache during trimming
Since commit 26583900327bf4cff3c8e9461d918889a103b317, we no longer
hold the global cache mutex throughout cache trimming, which
introduces a race condition when the currently-trimmed cache is
being disconnected in another thread, since we rely on the global
cache mutex for serialization.  Fix this, by acquiring the local
cache tile-storage mutex, in addition to the global cache mutex,
when disconnecting a cache.

Additionally, use one of the cache's fields to indicate that the
cache is being disconnected before acquiring any of the mutexes, so
that the trimming thread can quickly release them in response,
instead of keeping trimming the cache.