bin: add ability to clone parts of graph
This continus the use of similar or side-effect behavior of the
traditional copy paste keybindings in addition to control-x - which
already is used for removing nodes (it doesn't store them to a clipboard however).

now control-c stores a pointer to the active node and control-v uses the
pointer on the input port, or if aux port is active on aux port of a node, this
permits rigging up custom implementations of unsharp mask and drop-shadow from
primitive gegl nodes from scratch when using keybindings.