bin: permit connecting nodes using id= and ref= on commandline
This permits using the cursor keys to navigate the graph to a node, give it an
id with id=myid, and then navigate to a different node (possibly to aux port if
relevant) and issuing a ref=myid to create a link. Serialized the link will use
the id, while the actual in memory representation of GEGL is the connection
between the nodes, thus you do not need to rewire anything when a node is
renamed.

This commit also adds a guard against setting multiple nodes to the same id.