bin: add code to dispatch lua scripts
gboolean run_lua_file (const char *basename);

Tries running a lua sript for the GEGL ui by the name basename,
this is used by the code to look for on canvas ui code for operations
where : are replaced by _ in the op name. The code looks in the
following locations, using the first applicable.

/tmp/gegl_opname.lua
/if_running_out_of_source_tree/gegl/bin/lua/gegl_opname.lua
/home/user/.local/share/gegl-0.4/lua/gegl_opname.lua
/usr/share/gnome/gegl-0.4/lua/gegl_opname.lua
/prefix/share/gegl-0.4/lua/gegl_opname.lua
/usr/share/gegl-0.4/lua/gegl_opname.lua