bin: run some parts of ui from lua if right lua scripts exist
collection.lua and viewer.lua are the scripts looked for, if they
do not exist or lua support is not available the c versions are
run instead.

For now keybindings are registered for both c and lua, further
refactoring should make only a few global shortcuts be defined from
C leaving most to the lua side.