operations: add pdf-loading op
A simple page renderer that depends on poppler-glib, it render each page to a
linear buffer and becomes very slow for high ppi, default ppi set to 300 which
is sufficient to read and zoom in on visible details on most text based PDFs.

Maybe chunking/clipping the rendering with cairo would be better for high
resolution.