operations: propagate ffmpeg errors to GEGL.
This is actually the original issue I encountered when I realize GEGL
should report any unexpected error. In my case, we were exporting a MP4
video, which crashed the program. Apparently libx264 (and probably the
MP4 specification unless it's an implementation thing) does not allow
odd values for the dimensions. Now the crash is fixed with commit
6930afe68, yet the video export still fails. Even more wrong, since we
have no idea gegl_node_process() failed, and we run it for every frame
of our video, we just repeat the same error again and again.
Of course, the best is that the GUI forbids such values (now that I
know!), but the fact is that unexpected issues may always happen, and
when they do, it is nicer if the software can report them (currently,
ffmpeg errors are printed to stderr, thus only visible if you ran from
terminal).
So now the gegl:ff-save operation will propagate ffmpeg errors as a
GError (which can be later requested by gegl_node_process_success() upon
process failure).

Also updating the test-errors unit testing with such a failed MP4
export.