tests: improve unit tests for testing gegl_node_process_success().
load_denied() and save_denied() both failed on the CI infra. Instead of
having a "permission denied" error, we got an empty file error, which
means the test code was able to read the file. So the chmod must have
failed (probably system restrictions on the build environment).
So I transform the load_denied() to load_incomplete() with another type
of error and I skip the save_denied() if g_chmod() fails.