bin: add support for custom collection order and per file key/values
Add a filesystem based key/value store, and custom folder ordering that exist
in the per folder .gegl subfolder, which already contains the thumbnails.

For now there is no UI that permits creating a custom order, but creating
a file .gegl/index and having the filenames in the desired order, with
possible duplicates - will make gegl use the order provided in the file.