buffer: when building against glibc, issue malloc_trim after finalizing GeglBuffers
This will sometimes permit returning memory to the operating system when
just malloc/free calls alone do not do so. GeglBuffers are likely the largest
composite objects in a process with GEGL - thus here, or even periodically
in some other way is a way to force glibc's malloc/free to give up pages.