bin: add commandline interactions for setting/querying metadata
In collection view mode, the command foo=bar sets the key on the file foo to
bar, while foo@bar sets the attribute, which is associated with the entry in
the folder (which follows the file - but multiple entries can share a file).

The command info which in editing mode provides information about the selected
node now prints out the meta data associated with the selected file.