operation, transform: use indirect input buffer on main thread as well
In the various GeglOperation subclasses that use a per-thread
indirect input-buffer, backed by the actual input buffer, during
multithreaded processing, use an indirect buffer for the main
thread as well, instead of using the input buffer directly
(unless the main thread processes the entire output).

This avoids a deadlock when multiple threads gegl_buffer_copy()
from their input buffer to the output buffer concurrently: if a
worker thread acquires both the (indirect) input-buffer and output-
buffer locks, while the main thread acquires the (direct) input-
buffer lock and waits on the output-buffer lock, as soon as the
GeglTileBackendBuffer of the worker-thread's input buffer forwards
a command to the underlying input buffer, it will deadlock while
waiting for the input-buffer's lock (which is held by the main
thread).  See issue #157.

Furthermore, this should reduce contention over the underlying
input buffer between the main thread and the worker threads.