bin: add keyframe command
This is a command to be used when a specific property of an active
node is selected for editing, for now only doing linear interpolation
even though the GeglPath used underneath is a full bezier curve and
more interpolator.

 keyframe list
   prints the set keyframes
 keyframe set [prop] [value] [time]
   sets the currently set property value as a keyframe at current time
   overriding possible existing value
 keyframe toggle
   sets the currently set propery value as keyframe value at current time
   if doing so would be a no-op unset the existing value instead.
 keyframe clear
   removes animation curve of current property

The keybindings for "keyframe toggle" is control-k or just ` which is
easily accesible on english layouts.