workshop: add max iterations and cohesion parameters
The coehsion is implemented by taking the difference average of the
neighborhood of probes source neighborhood's average sampling location and
the probes sampling position.

Max iterations avoids infinite loops in the cases where pixels are not
being filled in at all.

Also expose indivdual ring radiuses, that when non 0 are taken directly
as the radius for that ring